server/requirements.txt
pycryptodome
Flask
elasticsearch
requests
tenacity